requirements.txt
telethon==1.40.0
Pillow>=11.0.0
aiofiles>=23.2.1
